+++ HISTORY.md
@@ -31,7 +31,7 @@
 * Enforce the existing contract of SingleDelete so that SingleDelete cannot be mixed with Delete because it leads to undefined behavior. Fix a number of unit tests that violate the contract but happen to pass.
 * ldb `--try_load_options` default to true if `--db` is specified and not creating a new DB, the user can still explicitly disable that by `--try_load_options=false` (or explicitly enable that by `--try_load_options`).
 * During Flush write or Compaction write/read, the WriteController is used to determine whether DB writes are stalled or slowed down. The priority (Env::IOPriority) can then be determined accordingly and be passed in IOOptions to the file system.
-
+* Use Env::IO_MID instead of Env::IO_LOW for the L0->L0 && L0->L1 compaction
 
 ## 7.2.0 (04/15/2022)
 ### Bug Fixes